+++ codemirror-ls/src/extensions/references.ts
@@ -33,7 +33,7 @@ import {
   showPanel,
 } from "@codemirror/view";
 import type * as LSP from "vscode-languageserver-protocol";
-import { LSNotSupportedError } from "../errors.js";
+import { LSNotSupportedError, NoReferencesError } from "../errors.js";
 import { LSCore } from "../LSPlugin.js";
 import { offsetToPos, posToOffset, posToOffsetOrZero } from "../utils.js";
 import type { LSExtensionGetter, Renderer } from "./types.js";
@@ -204,9 +204,7 @@ export async function handleFindReferences({
   const capability = REFERENCE_CAPABILITY_MAP[kind];
 
   if (!lsPlugin.client.capabilities?.[capability]) {
-    void lsPlugin._reportError(
-      new LSNotSupportedError("References not supported"),
-    );
+    lsPlugin._reportError(new LSNotSupportedError("References not supported"));
     return false;
   }
 
@@ -223,7 +221,7 @@ export async function handleFindReferences({
     });
 
     const onNoneFound = () =>
-      void lsPlugin._reportError(new NoReferencesError(kind));
+      lsPlugin._reportError(new NoReferencesError(kind));
 
     if (response === null) {
       onNoneFound();
@@ -247,7 +245,7 @@ export async function handleFindReferences({
     if (referenceLocations.length === 1 && goToIfOneOption) {
       const ref = referenceLocations[0];
       if (!ref) {
-        void lsPlugin._reportError(new NoReferencesError(kind));
+        lsPlugin._reportError(new NoReferencesError(kind));
         return false;
       }
 
@@ -401,12 +399,4 @@ function createReferencePanel(
       mount: () => panel.focus(),
     };
   };
-}
-
-export class NoReferencesError extends Error {
-  constructor(message: ReferenceKind) {
-    super(message);
-    this.name = "NoReferencesError";
-    this.message = `No ${REFERENCE_KIND_LABELS[message]} found`;
-  }
 }
